src/skills/builtin/calendar.py
"""Calendar management skill using Google Calendar API."""

from datetime import datetime, timezone
from typing import Any

from ..base import BaseSkill, SkillResult


class CalendarSkill(BaseSkill):
    """
    Skill for calendar management.

    Supports Google Calendar integration for reading
    and managing events.
    """

    name = "calendar"
    description = "Calendar management"
    version = "1.0.0"

    def __init__(self, config: dict[str, Any] | None = None) -> None:
        super().__init__(config)
        self.credentials_file = config.get("credentials_file")
        self._service: Any = None

    def _register_capabilities(self) -> None:
        """Register calendar capabilities."""
        self.register_capability(
            name="list_events",
            description="List calendar events",
            parameters={
                "type": "object",
                "properties": {
                    "calendar_id": {"type": "string", "default": "primary"},
                    "start_time": {"type": "string", "description": "Start time (ISO format)"},
                    "end_time": {"type": "string", "description": "End time (ISO format)"},
                    "max_results": {"type": "integer", "default": 10},
                },
            },
            security_action="calendar_read",
        )

        self.register_capability(
            name="create_event",
            description="Create a calendar event",
            parameters={
                "type": "object",
                "properties": {
                    "summary": {"type": "string", "description": "Event title"},
                    "start_time": {"type": "string", "description": "Start time (ISO format)"},
                    "end_time": {"type": "string", "description": "End time (ISO format)"},
                    "description": {"type": "string", "description": "Event description"},
                    "location": {"type": "string", "description": "Event location"},
                    "attendees": {"type": "array", "items": {"type": "string"}, "description": "Attendee emails"},
                    "calendar_id": {"type": "string", "default": "primary"},
                },
                "required": ["summary", "start_time", "end_time"],
            },
            security_action="calendar_write",
        )

        self.register_capability(
            name="update_event",
            description="Update an existing event",
            parameters={
                "type": "object",
                "properties": {
                    "event_id": {"type": "string", "description": "Event ID"},
                    "summary": {"type": "string", "description": "New title"},
                    "start_time": {"type": "string", "description": "New start time"},
                    "end_time": {"type": "string", "description": "New end time"},
                    "description": {"type": "string", "description": "New description"},
                    "calendar_id": {"type": "string", "default": "primary"},
                },
                "required": ["event_id"],
            },
            security_action="calendar_write",
        )

        self.register_capability(
            name="delete_event",
            description="Delete a calendar event",
            parameters={
                "type": "object",
                "properties": {
                    "event_id": {"type": "string", "description": "Event ID"},
                    "calendar_id": {"type": "string", "default": "primary"},
                },
                "required": ["event_id"],
            },
            security_action="calendar_write",
        )

        self.register_capability(
            name="find_free_time",
            description="Find free time slots",
            parameters={
                "type": "object",
                "properties": {
                    "start_time": {"type": "string", "description": "Search start (ISO format)"},
                    "end_time": {"type": "string", "description": "Search end (ISO format)"},
                    "duration_minutes": {"type": "integer", "description": "Required duration"},
                    "calendar_id": {"type": "string", "default": "primary"},
                },
                "required": ["start_time", "end_time", "duration_minutes"],
            },
            security_action="calendar_read",
        )

    async def initialize(self) -> None:
        """Initialize Google Calendar API."""
        if not self.credentials_file:
            self._initialized = False
            return

        try:
            from google.oauth2.credentials import Credentials
            from google_auth_oauthlib.flow import InstalledAppFlow
            from googleapiclient.discovery import build

            # TODO: Implement proper OAuth flow
            # For now, this is a placeholder
            self._initialized = False
        except ImportError:
            self._initialized = False

    async def execute(self, capability: str, **kwargs: Any) -> SkillResult:
        """Execute a calendar capability."""
        start_time = datetime.now(timezone.utc)

        if not self._initialized:
            return self._error_result(
                "Calendar not configured. Please set up Google Calendar API credentials.",
                start_time,
            )

        handlers = {
            "list_events": self._list_events,
            "create_event": self._create_event,
            "update_event": self._update_event,
            "delete_event": self._delete_event,
            "find_free_time": self._find_free_time,
        }

        handler = handlers.get(capability)
        if not handler:
            return self._error_result(f"Unknown capability: {capability}", start_time)

        try:
            result = await handler(**kwargs)
            return self._success_result(result, start_time)
        except Exception as e:
            return self._error_result(str(e), start_time)

    async def _list_events(
        self,
        calendar_id: str = "primary",
        start_time: str | None = None,
        end_time: str | None = None,
        max_results: int = 10,
    ) -> dict[str, Any]:
        """List calendar events."""
        # Placeholder implementation
        return {"events": [], "message": "Calendar not configured"}

    async def _create_event(
        self,
        summary: str,
        start_time: str,
        end_time: str,
        description: str | None = None,
        location: str | None = None,
        attendees: list[str] | None = None,
        calendar_id: str = "primary",
    ) -> dict[str, Any]:
        """Create a calendar event."""
        # Placeholder implementation
        return {"created": False, "message": "Calendar not configured"}

    async def _update_event(
        self,
        event_id: str,
        summary: str | None = None,
        start_time: str | None = None,
        end_time: str | None = None,
        description: str | None = None,
        calendar_id: str = "primary",
    ) -> dict[str, Any]:
        """Update a calendar event."""
        # Placeholder implementation
        return {"updated": False, "message": "Calendar not configured"}

    async def _delete_event(
        self,
        event_id: str,
        calendar_id: str = "primary",
    ) -> dict[str, Any]:
        """Delete a calendar event."""
        # Placeholder implementation
        return {"deleted": False, "message": "Calendar not configured"}

    async def _find_free_time(
        self,
        start_time: str,
        end_time: str,
        duration_minutes: int,
        calendar_id: str = "primary",
    ) -> dict[str, Any]:
        """Find free time slots."""
        # Placeholder implementation
        return {"slots": [], "message": "Calendar not configured"}
