pyproject.toml
[project]
name = "aria"
version = "0.1.0"
description = "Personal AI Agent - A fully-featured personal AI assistant"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Aria Team" }]
keywords = ["ai", "assistant", "agent", "llm", "automation"]

dependencies = [
    # Core
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",

    # Async & Queue
    "asyncio-mqtt>=0.16.0",
    "aiohttp>=3.9.0",
    "Brotli>=1.1.0",  # Required by aiohttp for Slack API brotli responses
    "redis>=5.0.0",
    "aioredis>=2.0.0",

    # LLM Providers
    "anthropic>=0.39.0",
    "ollama>=0.3.0",
    "openai>=1.50.0",  # For compatibility
    "tiktoken>=0.5.0",  # Token counting

    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",  # JWT
    "passlib[bcrypt]>=1.7.4",  # Password hashing

    # Messaging Channels
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.26.0",

    # Database & Storage
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "chromadb>=0.4.0",

    # Knowledge Graph
    "cognee>=0.5.0",

    # Browser Automation
    "playwright>=1.40.0",

    # Media Processing
    "edge-tts>=6.1.0",
    "openai-whisper>=20231117",
    "pillow>=10.0.0",
    "ffmpeg-python>=0.2.0",

    # Document Processing
    "reportlab>=4.0.0",
    "python-docx>=1.1.0",
    "pypdf>=3.17.0",

    # Communication
    "aiosmtplib>=3.0.0",
    "aioimaplib>=1.0.0",
    "twilio>=8.10.0",

    # Calendar
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "icalendar>=5.0.0",

    # Security & Crypto
    "cryptography>=41.0.0",
    "argon2-cffi>=23.1.0",

    # Utilities
    "httpx>=0.25.0",
    "structlog>=23.2.0",
    "rich>=13.7.0",
    "tenacity>=8.2.0",
    "orjson>=3.9.0",
    "pendulum>=3.0.0",

    # CLI Setup Wizard
    "questionary>=2.0.0",
    "qrcode>=7.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "black>=23.12.0",
    "pre-commit>=3.6.0",
    "httpx>=0.25.0",  # For testing FastAPI
]

[project.scripts]
aria = "src.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.ruff.isort]
known-first-party = ["aria"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
